requirement.txt
tensorflow==2.16.0
pillow==10.3.0
scikit-learn==1.4.2
google-cloud-storage>=2.16.0
google-cloud-aiplatform>=1.48.0
google-cloud-firestore>=2.16.0
google-cloud-vision>=3.7.2